pyproject.toml
[project]
name = "syndata-agent"
version = "0.1.0"
description = "A multi-agent framework for generating synthetic data using LLMS"
authors = [
    {name = "Pham Kinh Quoc", email = "phamkinhquoc2002@gmail.com"}
]
dependencies = [
    "python-dotenv==1.0.1",
    "typing-extensions==4.12.2",
    "pydantic==2.10.6",
    "openai==1.63.2",
    "google-genai==1.2.0",
    "ipykernel==6.29.5",
    "langchain==0.3.20",
    "langchain_community==0.3.19",
    "langgraph==0.2.74",
    "pytest==8.3.4",
    "pypdf==5.3.0",
    "faiss-cpu==1.10.0",
    "rich==13.9.4",
    "langchain-docling==0.2.0",
]

requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["syndata_agent"]